+++ task1_decoder_only_transformer/src/decoder_layer/single_decoder/local_single_self_decoder.py
@@ -53,10 +53,10 @@ class LocalSingleSelfDecoder(nn.Module):
       K = self.pe_model(K)
 
 
-    Q_chunks = [Q[:,j:j+w,:].float() for j in range(0,T,w)] #Q_chunk[0] : (B,w,d_k)
-    K_chunks = [K[:,j:j+w,:].float() for j in range(0,T,w)]  #K_chunks.T(-2,-1) : (B,d_k,w) 
+    Q_chunks = [Q[:,j:j+w,:] for j in range(0,T,w)] #Q_chunk[0] : (B,w,d_k)
+    K_chunks = [K[:,j:j+w,:] for j in range(0,T,w)]  #K_chunks.T(-2,-1) : (B,d_k,w) 
     # @ = ( B,w,d_k) @ (B,d_k,w) -> (B,w,w) @(B,w,d_k) -> (B,w,d_k)
-    V_chunks = [V[:,j:j+w,:].float() for j in range(0,T,w)]
+    V_chunks = [V[:,j:j+w,:] for j in range(0,T,w)]
 
     chunk_curr = torch.stack([Q_chunks[i]@K_chunks[i].transpose(-2,-1) for i in range(len(Q_chunks))])
